Refactored, disassembled and improved CreateAssessment component
